move SplineInterpolationPoints2d to another file
